Clean the project and rename the namespace to pascal-case
