+++ csharp-basic-topics/ZipFilesInNet/Tests/ReadZipFilesTest.cs
@@ -1,4 +1,4 @@
-﻿using Zip_files_in_NET;
+﻿using ZipFilesInNet;
 
 namespace Tests
 {
