+++ csharp-basic-topics/ZipFilesInNet/ZipFilesInNET/ModifyZipFiles.cs
@@ -1,7 +1,7 @@
 ﻿using System.IO.Compression;
 using System.Text.RegularExpressions;
 
-namespace Zip_files_in_NET
+namespace ZipFilesInNet
 {
     public class ModifyZipFiles
     {
