+++ csharp-basic-topics/ZipFilesInNet/ZipFilesInNET/ZipFilesInNET.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>Zip_files_in_NET</RootNamespace>
+    <RootNamespace>ZipFilesInNet</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
